task43

+++ seminar6/HW_task40,41,43/task43/Program.cs
@@ -5,13 +5,13 @@
 // b1 = 2, k1 = 5, b2 = 4, k2 = 9 -> (-0,5; -0,5)
 
 Console.WriteLine("Введите значение b1:");
-int b1 = Convert.ToInt32(Console.ReadLine());
+double b1 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine("Введите значение k1:");
-int k1 = Convert.ToInt32(Console.ReadLine());
+double k1 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine("Введите значение b2:");
-int b2 = Convert.ToInt32(Console.ReadLine());
+double b2 = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine("Введите значение k2:");
-int k2 = Convert.ToInt32(Console.ReadLine());
+double k2 = Convert.ToInt32(Console.ReadLine());
 
 if(k1==k2 && b1==b2)
 {
@@ -21,8 +21,8 @@ else
 {
     if(k1 != k2)
     {
-        int x = (b2-b1)/(k1-k2);
-        int y = k1*x +b1;
+        double x = (b2-b1)/(k1-k2);
+        double y = k1*x +b1;
         Console.WriteLine($"Координата точки пересечения прямых: {x}, {y}");
         }
         else
